examples/fastapi-simple/requirements.txt
haredis
fastapi[all]